requirements.txt
crewai
langchain
langchain-anthropic
fastapi
uvicorn
celery
redis
apscheduler
sqlalchemy
alembic
asyncpg
psycopg2-binary
pydantic
python-dotenv
httpx
python-telegram-bot
pillow
boto3
sentry-sdk
streamlit
plotly
loguru
